Stars as component with custom colors
